+++ doc/changes.rst
@@ -5,7 +5,13 @@ desitarget Change Log
 0.19.2 (unreleased)
 -------------------
 
-* No changes yet.
+* Update of sky selection code [`PR #290`_]. Includes:
+   * Use the :mod:`desitarget.brightmask` formalism to speed calculations.
+   * Pass around a magnitude limit on masks from the sweeps (to better
+     avoid only objects that are genuinely detected in the sweeps).
+   * Reduce the default margin to produce ~1700 sky positions per sq. deg.
+   
+.. _`PR #290`: https://github.com/desihub/desitarget/pull/290
 
 0.19.1 (2018-03-01)
 -------------------
@@ -140,7 +146,7 @@ to importing target mask bits and how mock spectra are generated.
 0.14.0 (2017-07-10)
 -------------------
 
-* Significant updated to handle transition from pre-DR4 to post-DR4 data model [`PR #189`_]:
+* Significant update to handle transition from pre-DR4 to post-DR4 data model [`PR #189`_]:
    * :mod:`desitarget.io` can now read old DR3-style and new DR4-style tractor and sweeps files
    * :mod:`desitarget.cuts` now always uses DR4-style column names and formats
    * new 60-bit ``TARGETID`` schema that incorporates ``RELEASE`` column from imaging surveys